+++ tests/test_import_dbt.py
@@ -84,12 +84,14 @@ models:
       order_id:
         type: integer
         description: This is a unique identifier for an order
+        primaryKey: true
         required: true
         unique: true
       customer_id:
         type: integer
         description: Foreign key to the customers table
         required: true
+        references: customers.customer_id
       order_date:
         type: date
         description: Date (UTC) that the order was placed
@@ -141,6 +143,7 @@ models:
       customer_id:
         type: integer
         description: ''
+        primaryKey: true
         required: true
         unique: true
       first_name:
@@ -156,6 +159,7 @@ models:
       order_id:
         type: integer
         description: ''
+        primaryKey: true
         required: true
         unique: true
       customer_id:
@@ -174,6 +178,7 @@ models:
       payment_id:
         type: integer
         description: ''
+        primaryKey: true
         required: true
         unique: true
       order_id:
@@ -193,6 +198,7 @@ models:
       customer_id:
         type: integer
         description: This is a unique identifier for a customer
+        primaryKey: true
         required: true
         unique: true
       first_name:
@@ -242,12 +248,14 @@ models:
       order_id:
         type: bigint
         description: This is a unique identifier for an order
+        primaryKey: true
         required: true
         unique: true
       customer_id:
         type: bigint
         description: Foreign key to the customers table
         required: true
+        references: customers.customer_id
       order_date:
         type: date
         description: Date (UTC) that the order was placed
@@ -299,6 +307,7 @@ models:
       customer_id:
         type: bigint
         description: ''
+        primaryKey: true
         required: true
         unique: true
       first_name:
@@ -314,6 +323,7 @@ models:
       order_id:
         type: bigint
         description: ''
+        primaryKey: true
         required: true
         unique: true
       customer_id:
@@ -332,6 +342,7 @@ models:
       payment_id:
         type: bigint
         description: ''
+        primaryKey: true
         required: true
         unique: true
       order_id:
@@ -351,6 +362,7 @@ models:
       customer_id:
         type: bigint
         description: This is a unique identifier for a customer
+        primaryKey: true
         required: true
         unique: true
       first_name:
@@ -422,6 +434,7 @@ models:
       customer_id:
         type: integer
         description: This is a unique identifier for a customer
+        primaryKey: true
         required: true
         unique: true
       first_name: